Moved some remaining parts to derivations package
