+++ engine-core/src/main/scala/org/scalarules/derivations/DerivationGraph.scala
@@ -1,7 +1,5 @@
 package org.scalarules.derivations
 
-import org.scalarules.engine.Derivation
-
 trait DerivationGraph {
 
 }
