+++ engine-core/src/main/scala/org/scalarules/derivations/derivations.scala
@@ -1,10 +1,12 @@
-package org.scalarules.engine
+package org.scalarules.derivations
 
 // TODO : Turn this off and fix it
-import scala.language.existentials
-import DerivationTools._
+import org.scalarules.derivations.DerivationTools._
+import org.scalarules.engine._
 import org.scalarules.facts.Fact
 import org.scalarules.utils.{SourcePosition, SourceUnknown}
+
+import scala.language.existentials
 
 /**
   * A `Derivation` models the basic element used inside the `Engine`. One execution step of the `Engine` will always
