+++ engine-core/src/main/scala/org/scalarules/engine/metadata.scala
@@ -1,4 +1,6 @@
 package org.scalarules.engine
 
+import org.scalarules.derivations.Derivation
+
 case class Step(initial: Context, derivation: Derivation, status: String, result: Context)
