+++ engine-core/src/main/scala/org/scalarules/utils/PrettyPrinter.scala
@@ -1,9 +1,10 @@
 package org.scalarules.utils
 
+import org.scalarules.derivations.{DefaultDerivation, Derivation, SubRunDerivation}
 import org.scalarules.engine._
 
 import scala.annotation.tailrec
-import org.scalarules.engine.DerivationTools._
+import org.scalarules.derivations.DerivationTools._
 import org.scalarules.facts.Fact
 
 object PrettyPrinter {
