+++ engine-test-utils/src/main/scala/org/scalarules/testutils/TestUtils.scala
@@ -1,6 +1,7 @@
 package org.scalarules.testutils.nl
 
-import org.scalarules.engine.{Derivation, FactEngine, Step, _}
+import org.scalarules.derivations.Derivation
+import org.scalarules.engine.{FactEngine, Step, _}
 import org.scalarules.facts.Fact
 import org.scalarules.services.{AnalyzedScenario, HeuristicService}
 import org.scalarules.utils.PrettyPrinter
